Temporarily stop using a utils module
We are having some difficulties on how to reference to a shared utils module.
Temporarily move the shared code to each caller module util a proper fix.
