+++ plugins/event_source/aws_cloudtrail.py
@@ -37,7 +37,6 @@ from datetime import datetime
 from typing import Any, Dict
 
 from aiobotocore.session import get_session
-from ansible_collections.ansible.eda.plugins.utils.aws_utils import connection_args
 
 
 def _cloudtrail_event_to_dict(event):
@@ -102,6 +101,25 @@ async def main(queue: asyncio.Queue, args: Dict[str, Any]):
             event_time = c_event_time or event_time
 
             await asyncio.sleep(delay)
+
+
+def connection_args(args: Dict[str, Any]) -> Dict[str, Any]:
+    selected_args = {}
+
+    # Best Practice: get credentials from ~/.aws/credentials or the environment
+    # However the following method are still possible
+    if "access_key" in args:
+        selected_args["aws_access_key_id"] = args["access_key"]
+    if "secret_key" in args:
+        selected_args["aws_secret_access_key"] = args["secret_key"]
+    if "session_token" in args:
+        selected_args["aws_session_token"] = args["session_token"]
+
+    if "endpoint_url" in args:
+        selected_args["endpoint_url"] = args["endpoint_url"]
+    if "region" in args:
+        selected_args["region_name"] = args["region"]
+    return selected_args
 
 
 if __name__ == "__main__":
