--- plugins/utils/aws_utils.py
@@ -1,20 +1,0 @@
-from typing import Any, Dict
-
-
-def connection_args(args: Dict[str, Any]) -> Dict[str, Any]:
-    selected_args = {}
-
-    # Best Practice: get credentials from ~/.aws/credentials or the environment
-    # However the following method are still possible
-    if "access_key" in args:
-        selected_args["aws_access_key_id"] = args["access_key"]
-    if "secret_key" in args:
-        selected_args["aws_secret_access_key"] = args["secret_key"]
-    if "session_token" in args:
-        selected_args["aws_session_token"] = args["session_token"]
-
-    if "endpoint_url" in args:
-        selected_args["endpoint_url"] = args["endpoint_url"]
-    if "region" in args:
-        selected_args["region_name"] = args["region"]
-    return selected_args
