+++ tests/unit/event_source/test_aws_cloudtrail.py
@@ -2,11 +2,10 @@ import datetime
 from unittest.mock import patch
 
 import pytest
-from ansible_collections.ansible.eda.plugins.event_source.aws_cloudtrail import (
-    main as cloudtrail_main,
-)
 from asyncmock import AsyncMock
 from mock import MagicMock
+
+from plugins.event_source.aws_cloudtrail import main as cloudtrail_main
 
 
 @pytest.mark.asyncio
@@ -29,7 +28,7 @@ async def test_receive_from_cloudtrail(eda_queue):
         ]
     }
     with patch(
-        "ansible_collections.ansible.eda.plugins.event_source.aws_cloudtrail.get_session",  # noqa: E501
+        "plugins.event_source.aws_cloudtrail.get_session",  # noqa: E501
         return_value=session,
     ):
         iterator = AsyncMock()
